refactor: `connect` Menu component instead of Navigator component
